==> notebooks/1.1-dataset_statistics.ipynb <==
# Dataset statistics

This notebook includes the statistics analysis for each dataset

# Imports

import os
import json
import pandas as pd

from utils import DATA_DIR

# Load datasets

with open(os.path.join(DATA_DIR, 'transcriptomics', 'creed_harmonized_expression.json')) as f:
    creed = json.load(f)
    
with open(os.path.join(DATA_DIR, 'transcriptomics', 'geo_harmonized_expression.json')) as f:
    geo = json.load(f)
    
with open(os.path.join(DATA_DIR, 'transcriptomics', 'l1000_harmonized_expression.json')) as f:
    l1000 = json.load(f)
    
with open(os.path.join(DATA_DIR, 'transcriptomics', 'target_harmonized_expression.json')) as f:
    target = json.load(f)

MAP = {
    'creed': creed,
    'target': target,
    'geo': geo,
    'l1000': l1000
}

# Statistical analysis

df = pd.DataFrame(columns=[
    'dataset',
    'count', 
    'mean_gene_count', 
    'minimum_gene_count',
    'maximum_gene_count'
])

for dataset_name, data in MAP.items():
    
    data_count = len(data)
    
    gene_list = []
    
    for el in data:
        gene_list.append(len(data[el]))
    
    max_count = max(gene_list)
    min_count = min(gene_list)
    avg_count = sum(gene_list) / len(gene_list)
    
    tmp = pd.DataFrame(
        {'dataset': dataset_name,
        'count': data_count, 
        'mean_gene_count': avg_count, 
        'minimum_gene_count': min_count,
        'maximum_gene_count': max_count},
        index=[1,]
    )
    
    df = pd.concat([df, tmp], ignore_index=True)

df
