==> notebooks/2.0-dataset_probabilities.ipynb <==
# Dataset probabilities

This notebook involves analysing the clinical trial information for chemical-disease pair in for a given dataset to get the probability by chance.

# Imports

import os
import logging
import json
from collections import Counter
from tqdm import tqdm
import pandas as pd
from itertools import product

from networkx import DiGraph
from networkx.algorithms.shortest_paths.generic import has_path

from utils import KG_DATA_PATH, DATA_DIR, create_graph_from_df

logger = logging.getLogger(__name__)

# Load KG

openbiolinks_df = pd.read_csv(
    os.path.join(KG_DATA_PATH, 'openbiolink_filtered_kg.tsv'),
    sep='\t'
)

custom_df = pd.read_csv(
    os.path.join(KG_DATA_PATH, 'custom_filtered_kg.tsv'),
    sep='\t'
)

openbio_graph = create_graph_from_df(openbiolinks_df)
custom_graph = create_graph_from_df(custom_df)

# Loading clinical trial pairs

with open(os.path.join(DATA_DIR, 'gold-standard', 'clinical-trial.json'), 'r') as f:
    clinical_filtered = json.load(f)

# GEO, LC1000, OpenTargets and Creed

# Load gene expression files

with open(os.path.join(DATA_DIR, 'transcriptomics', 'creed_harmonized_expression.json')) as file:
    creed_chemicals = json.load(file).keys()
    
with open(os.path.join(DATA_DIR, 'transcriptomics', 'geo_harmonized_expression.json')) as file2:
    geo_diseases = json.load(file2).keys()
    
with open(os.path.join(DATA_DIR, 'transcriptomics', 'lc1000_harmonized_expression.json')) as file3:
    lc1000_chemicals = json.load(file3).keys()
    
with open(os.path.join(DATA_DIR, 'transcriptomics', 'target_harmonized_expression.json')) as file4:
    open_target_diseases = json.load(file4).keys()


# Get overlaps with subgraph

MAP = {
    'creed' : creed_chemicals,
    'target': open_target_diseases,
    'geo': geo_diseases,
    'lc1000': lc1000_chemicals,
}

def get_overlap_stats(graph_dict: dict):
    
    openbio_graph = graph_dict['openbio']
    
    openlink_kg_chemical = set()
    openlink_kg_disease = set()

    for node in openbio_graph.nodes():
        if 'pubchem' in node:
            openlink_kg_chemical.add(node)

        elif 'mondo' in node:
            openlink_kg_disease.add(node)

    custom_graph = graph_dict['custom']
    
    custom_kg_chemical = set()
    custom_kg_disease = set()

    for node in custom_graph.nodes():
        if 'pubchem' in node:
            custom_kg_chemical.add(node)

        elif 'mondo' in node:
            custom_kg_disease.add(node)
            
    # Get data present on in the KG 
    info_dict = {}

    for i in MAP:
        curies = set(MAP[i])

        print(f'\n#### {i} ####')
        disease1 = openlink_kg_disease.intersection(curies)
        disease2 = custom_kg_disease.intersection(curies)

        chemical1 = openlink_kg_chemical.intersection(curies)
        chemical2 = custom_kg_chemical.intersection(curies)

        if i not in info_dict:
            info_dict[i] = {}

        if len(disease1) > 0:
            info_dict[i]['openbiolinks'] = disease1
            info_dict[i]['custom'] = disease2
        elif len(chemical1) > 0:
            info_dict[i]['openbiolinks'] = chemical1
            info_dict[i]['custom'] = chemical2
        else:
            print('SOMETHING IS WRONG!!!')

        print(f'Overlaps with OpenBioLink KG chemical', len(chemical1))
        print(f'Overlaps with Custom KG chemical', len(chemical2))
        print(f'Overlaps with OpenBioLink KG disease', len(disease1))
        print(f'Overlaps with Custom KG disease', len(disease2))
    return info_dict

info_graph = get_overlap_stats(graph_dict={'openbio': openbio_graph, 'custom': custom_graph})

# Get clinical pair percentage based on dataset and gold standard data

def get_gold_standard_stats(
    gold_standard_dict: dict, 
    graph_dict: dict, 
    info_dict: dict
):
    
    clinical_evidence = {'openbiolinks': {}, 'custom': {}}
    count_dict = {'openbiolinks': {}, 'custom': {}}
    total_count_dict = {'openbiolinks': {}, 'custom': {}}
    
    openbio_graph = graph_dict['openbio']
    custom_graph = graph_dict['custom']

    for c, d in product(['creed', 'lc1000'], ['target', 'geo']):
        c_set = info_dict[c]
        d_set = info_dict[d]
        
        # For OpenBioLinks
        openbio_pairs_with_paths = 0
        tp_openbio = 0
        
        for chem_idx, disease_idx in product(c_set['openbiolinks'], d_set['openbiolinks']):
            if has_path(G=openbio_graph, source=chem_idx, target=disease_idx):
                openbio_pairs_with_paths += 1
                name = chem_idx + '_' + disease_idx
                
                if name in gold_standard_dict:
                    clinical_evidence['openbiolinks'][name] = 'yes'
                    tp_openbio += 1

        key_name = c + '_' + d
        
        # No.of clinical trial pairs
        count_dict['openbiolinks'][key_name] = tp_openbio  
        
        # Total no.of pairs
        total_count_dict['openbiolinks'][key_name] = openbio_pairs_with_paths
        
        # For Custom Network
        custom_pairs_with_paths = 0
        tp_custom = 0

        for chem_idx, disease_idx in product(c_set['custom'], d_set['custom']):
            if has_path(G=custom_graph, source=chem_idx, target=disease_idx):
                custom_pairs_with_paths += 1
                name = chem_idx + '_' + disease_idx
                if name in gold_standard_dict:
                    clinical_evidence['custom'][name] = 'yes'
                    tp_custom += 1

        key_name = c + '_' + d
        
        # No.of clinical trial pairs
        count_dict['custom'][key_name] = tp_custom
        
        # Total no.of pairs
        total_count_dict['custom'][key_name] = custom_pairs_with_paths
    
    # Display the value by change df
    data_df = pd.DataFrame(
        columns=['creed_target', 'creed_geo', 'lc1000_target', 'lc1000_geo'],
        index=['openbiolinks', 'custom']
    )    
    
    for kg_name in total_count_dict:
        for el in total_count_dict[kg_name]: 
            val = count_dict[kg_name][el] / total_count_dict[kg_name][el]
            data_df.loc[kg_name, el] = val*100
    
    return data_df, count_dict

# Clinical data - Value by chance

df, counts = get_gold_standard_stats(
    clinical_filtered, 
    graph_dict={'openbio': openbio_graph, 'custom': custom_graph},
    info_dict=info_graph,
)

df

counts
