==> notebooks/3.0-kg_statistics.ipynb <==
# Kg-statistics

This notebook compares the graph statistics of the permuted and the raw graphs

# Imports

import os
import pandas as pd
import networkx as nx

from utils import KG_DATA_PATH, create_graph_from_df

# Helper Functions

def get_longest_shortest_path(df, graph):
    longest = 5e-324
    
    drugs = set(df[df['source'].str.startswith('pubchem')]['source'])
    diseases = set(df[df['target'].str.startswith('mondo')]['target'])
    
    for drug in drugs:
        for disease in diseases:
            # some paths do not exist
            try:
                path_length = len(nx.shortest_path(graph, source = drug, target = disease))
            except:
                continue
            if path_length > longest:
                longest = path_length
    
    return longest

# Load data

custom = pd.read_csv(
    os.path.join(KG_DATA_PATH, 'custom_filtered_kg.tsv'), 
    sep='\t'
)

openbiolink = pd.read_csv(
    os.path.join(KG_DATA_PATH, 'openbiolink_filtered_kg.tsv'), 
    sep='\t'
)
openbiolink

# Create graph

graph_openbio = create_graph_from_df(openbiolink)
graph_custom = create_graph_from_df(custom)

# Graph stats

data = {
    'openbio': {},
    'custom': {}
}

openbio_g = graph_openbio
custom_g = graph_custom
openbio_df = openbiolink
custom_df = custom


o_nodes = list(openbio_g.nodes())

o_degree = list(openbio_g.degree(o_nodes))
o_degree = [el for node, el in o_degree]

c_nodes = list(custom_g.nodes())

c_degree = list(custom_g.degree(c_nodes))
c_degree = [el for node, el in c_degree]


data['openbio']['# nodes'] = str(len(o_nodes))
data['custom']['# nodes'] = str(len(c_nodes))

data['openbio']['# degree'] = str(round(sum(o_degree) / len(o_degree), 3))
data['custom']['# degree'] = str(round(sum(c_degree) / len(c_degree), 3))

o_edges = len(openbio_g.edges())
c_edges = len(custom_g.edges())

data['openbio']['# edges'] = str(o_edges)
data['custom']['# edges'] = str(c_edges)

data['openbio']['# activatory'] = len(openbio_df.index[openbio_df['polarity'] == 1])
data['custom']['# activatory'] = len(custom_df.index[custom_df['polarity'] == 1])

data['openbio']['# inhibitory'] = len(openbio_df.index[openbio_df['polarity'] == -1])
data['custom']['# inhibitory'] = len(custom_df.index[custom_df['polarity'] == -1])

data['openbio']['longest_shortest_path'] = get_longest_shortest_path(openbio_df, openbio_g)
data['custom']['longest_shortest_path'] = get_longest_shortest_path(custom_df, custom_g)    

df = pd.DataFrame(data)
df
