==> notebooks/5.0-benchmark_analysis.ipynb <==
# Benchmark analysis

This notebook compares the different benchmark according to the two KG.

# Imports

import os
import json
import math
import logging
import networkx as nx
import numpy as np
import pandas as pd

from collections import Counter, defaultdict
from itertools import product
from typing import Mapping, List
from tqdm import tqdm

from utils import DATA_DIR, KG_DATA_PATH, create_graph_from_df

# Load graph

openbiolink_df = pd.read_csv(
    os.path.join(KG_DATA_PATH, 'openbiolink_filtered_kg.tsv'),
    sep='\t'
)

custom_df = pd.read_csv(
    os.path.join(KG_DATA_PATH, 'custom_filtered_kg.tsv'),
    sep='\t'
)


# Load clinical data

with open(os.path.join(DATA_DIR, 'gold-standard', 'clinical-trial.json')) as file:
    clinical_pair_dict = json.load(file).keys()

# Create KG 

graph_openbio = create_graph_from_df(openbiolink_df)
graph_custom = create_graph_from_df(custom_df)

# Create drug disease dicts specific to KG

disease_openbio, drug_openbio,disease_custom,drug_custom = set(), set(), set(), set()

for u in graph_openbio.nodes():
    if u.startswith('mondo'):
        disease_openbio.add(u)
    elif u.startswith('pubchem'):
        drug_openbio.add(u)

for u in graph_custom.nodes():
    if u.startswith('mondo'):
        disease_custom.add(u)
    elif u.startswith('pubchem'):
        drug_custom.add(u)

# Helper Functions

score_actual = {}
kg_dfs = {}

def khop(
    nodeA: str, 
    nodeB: str, 
    graph: nx.Graph, 
    total: bool
) -> tuple:
    
    """Find nodes within the distance limit """
    
    khop_A = {u for u in graph.neighbors(nodeA)}
    khop_B = {u for u in graph.neighbors(nodeB)}
    
    if total:
        return list(khop_A | khop_B), khop_A, khop_B
    else:
        return list(khop_A & khop_B), khop_A, khop_B

def get_dict_df(
    diseases, 
    drugs, 
    undirected_kg_graph, 
    di_kg_graph,
    similarity_type
):
    
    df = pd.DataFrame(columns=[
        'source',
        'target',
        sim_scores[similarity_type]
    ])

    for disease in diseases:
        # Skip drugs not part of largest component of KG
        if disease not in undirected_kg_graph.nodes():
            continue
        
        cn = []
        
        # for each disease, find the similarity score with for each drug and append to list
        for drug in drugs:
            
            # Skip drugs not part of largest component of KG
            if drug not in undirected_kg_graph.nodes():
                continue
            
            shared_nodes, nodeA_neighbor, nodeB_neighbor = khop(
                nodeA=drug,
                nodeB=disease,
                graph=undirected_kg_graph, 
                total=False,
            )
            
            total_nodes, _, _ = khop(
                nodeA=drug,
                nodeB=disease,
                graph=undirected_kg_graph, 
                total=True,
            )
                
            if similarity_type == 'cn':
                similarity = len(shared_nodes)
            
            elif similarity_type == 'cos':
                similarity = len(shared_nodes) / (math.sqrt(len(nodeA_neighbor) * len(nodeB_neighbor)))
            
            elif similarity_type == 'ji':
                similarity = len(shared_nodes) / len(total_nodes)
                
            elif similarity_type == 'si':
                similarity = (2 * len(shared_nodes)) / (len(nodeA_neighbor) + len(nodeB_neighbor))
                
            elif similarity_type == 'hpi':
                similarity = len(shared_nodes) / min(len(nodeA_neighbor), len(nodeB_neighbor)) 
                
            elif similarity_type == 'hdi':
                similarity = len(shared_nodes) / max(len(nodeA_neighbor), len(nodeB_neighbor)) 
                
            elif similarity_type == 'lhn':
                similarity = len(shared_nodes) / (len(nodeA_neighbor) * len(nodeB_neighbor)) 
                
            elif similarity_type == 'pa':
                similarity = len(nodeA_neighbor) * len(nodeB_neighbor)
                
            elif similarity_type == 'aa':
                similarity = 0
                
                for n in shared_nodes:
                    neighbors_list = set(i for i in undirected_kg_graph.neighbors(n))
                    similarity += 1 / math.log10(len(neighbors_list))
                    
            elif similarity_type == 'ra':
                similarity = 0
                
                for n in shared_nodes:
                    neighbors_list = set(i for i in undirected_kg_graph.neighbors(n))
                    similarity += 1 / len(neighbors_list)
            elif similarity_type == 'sp':
                # try to see if path is between two nodes
                try:
                    similarity = len(nx.shortest_path(di_kg_graph,source=drug, target=disease))
                except:
                    continue
            cn.append(similarity)
        
        if not similarity_type == 'sp':
            index = np.where(cn == np.amax(cn))
        else:
            index = np.where(cn == np.amin(cn))
        # if list is full of 0's (i.e sum == 0), then there are no shared neighbors 
        
        if np.sum(cn) == 0: #or len(np.array(cn)[index]) > 1:
            continue   
        
        for val in index:
            for j in val:
                df = df.append(
                    {
                        'source': list(drugs)[j], 
                        'target': disease, 
                        sim_scores[similarity_type]: cn[j]
                    }, 
                    ignore_index=True
                )

    return df

def get_precision(
    clinical_trial_dict: dict, 
    predicted: list
)-> tuple: 
    
    total = len(predicted)
    pos = 0
    
    for pair in predicted:
        if pair in clinical_trial_dict:
            pos += 1
    
    
    return round(((pos/total) * 100), 3), pos, total


# Value by change for both KGs

openbio_prob = 0

for disease in disease_openbio:
    for drug in drug_openbio:
        trial = f'{drug}_{disease}'
        if trial in clinical_pair_dict:
            openbio_prob += 1

total = len(drug_openbio) * len(disease_openbio)
prob = openbio_prob / total
open_val_by_chance = round(prob, 3)
open_val_by_chance

custom_prob = 0

for disease in disease_custom:
    for drug in drug_custom:
        trial = f'{drug}_{disease}'
        if trial in clinical_pair_dict:
            custom_prob += 1

total = len(drug_custom) * len(disease_custom)
prob = custom_prob / total
custom_val_by_chance = round(prob, 3)
custom_val_by_chance

# Different benchmark methods

sim_scores = {
    'cn': 'Common Neighbors',
    'cos': 'Cosine Similiarity',
    'ji': 'Jaccard index',
    'si': 'Sorensen index',
    'hpi': 'Hub Promoted Index',
    'hdi': 'Hub Depressed Index', 
    'lhn': 'Leicht–Holme–Newman Index',
    'pa':'Preferential Attachment',
    'aa': 'Adamic-Adar', 
    'ra': 'Resource Allocation Index',
    'sp': 'Shortest Path'
}

undirected_openbio = graph_openbio.to_undirected()
undirected_custom = graph_custom.to_undirected()
check_df = None

for algo in tqdm(sim_scores, desc='Calculating scores for algorithms'):
    
    algo_name = sim_scores[algo]
    
    # OpenBioLink KG
    full_df = get_dict_df(
        diseases=disease_openbio,
        drugs=drug_openbio, 
        undirected_kg_graph=undirected_openbio,
        di_kg_graph=graph_openbio,
        similarity_type=algo
    )

        
    full_df['pair'] = full_df['source'] + '_' + full_df['target']
    
    openbio_precision, openbio_pos, openbio_total = get_precision(
        clinical_trial_dict=clinical_pair_dict, 
        predicted=list(full_df['pair'].unique())
    )
    
    # Custom KG
    df_custom = get_dict_df(
        diseases=disease_custom, 
        drugs=drug_custom, 
        undirected_kg_graph=undirected_custom,
        di_kg_graph=graph_custom,
        similarity_type=algo
    )
    
    df_custom['pair'] = df_custom['source'] + '_' + df_custom['target']
    
    custom_precision, custom_pos, custom_total = get_precision(
        clinical_trial_dict=clinical_pair_dict, 
        predicted=list(df_custom['pair'].unique())
    )
   
    score_actual[algo_name] = {
        'openbio_precision': openbio_precision,
        'openbio_val_by_chance': open_val_by_chance,
        '# openbio_pairs': f'{openbio_pos}/{openbio_total}',
        'custom_precision': custom_precision,
        'custom_val_by_chance': custom_val_by_chance,
        '# custom_pairs': f'{custom_pos}/{custom_total}',
    }    

pd.DataFrame(score_actual).transpose()
